Initial Commit
This is a set of scripts to run molecular dynamics simulations using
Hoomd. Currently is a set of scripts to create  equilibrated
files at a series of temperatures.

There is currently some issues with the equilibration, in that the
parameters fluctuate over a large range and the equilibration parameters
do not match the input parameters.