Documenting using google style
Converting all the documentation to the google stye of documentation for
consistency. This is the style of documentation that I prefer over other
implementations.

I have also updated the documentation to include a whole bunch of useful
information like variable types of input/output variables and equations
for the dynamics quantities.