Class to read and write data to a JSON dictionary
All the dynamics data can be computed only knowing the translation and
rotation of each molecule along with the time difference between the
initial and final configurations. I have developed a class to read and
write this data to a JSON file, which will allow me to develop new
analyses or redo the analysis at a later time.