Restructuring dynamics calculations
Breaking the computation of the dynamics into 3 distinct components to
make it easier to deal with a workflow where the data can either come
from running hoomd or from a file.

The TransData module is to hold all the translational, rotational and
timestep data. It deals with the input and output of the data to files.

The CompDynamics modules does all the computation. It can also print the
resulting computations (as well as rotational and translational data) to
a file. The data is internally stored as a TransData object.

The TimeDep module interfaces the CompDynamics module with hoomd,
performing the calculations for the deltas in position and rotation.