Hoomd v2 support
Converting scripts running the simulations and equilibrations to run
using the updated Hoomd 2.x interface. This has required a significant
number of changes.

Most notable of the changes is that file format, changing from xml to
gsd files. This is because gsd is now the default for hoomd 2.x. While
there is little support for this format, it should grow relatively
quickly and also should last for a significant period of time.

Also changed is how the system deals with rigid bodies, these now have
to be created and initialised. This isn't such an issue, but is a
problem with compatability to hoomd 1.x data since the central particles
are not contigous in number.

Another reasonably significant change is that the initial configuration
is no longer a pre generated crystal. The rigid bodies are formed from
the `create_rigid()` command and so we start with the central bodies in
a hexagonal lattice. This will make the generalisation of this script to
differnt molecuels much simpler.