Updating TimeDep module for hoomdv2
Updating the analysis interface with hoomd to be compatable with the new
API. There are many benefits to the new interface, primarily the
inclusing of the orientation in the snapshot data. Also convenient is
there is no longer a difference between single particle data and
molecular data apart from only counting certain particles.