Correcting moments of inertia
Only need the moment of inertia for the z direction, which corresponds
to that on the x,y plane on which my molecules reside. This corrects
that moment of inertia changing it from 10, to 1.65 which it should be.

Setting the moment of inertia to a huge value may be one way to restrict
the rotations of the moelcules.