Output and read GSD files
Outputing GSD files for the configuration data allowsing later analysis
of the simulation. This will hopefully be more efficient than my current
plan of outputing translations and rotations.

Unfortunately there are a few differences between the snapshots in the
GSD files and the snapshots in hoomd, however almost all of the
analysis when reading the GSD files is the same as when dealing with
snapshots during the simulation run.