Performance - Removing decoupling from print_all
Removing the decoupling parameter from the print_all functions for the
CompDynamics module. This was the main contributor to the slow
computations and since I don't actually use it anyway it was a simple
question of removing that computation. There is a vast performance
improvement, on the order of hundreds of times.

I have also reconfigured the if statement when adding key frames to
better deal with the current way I am collecting the frames.