Bugfix - Fixing the keyframes conputed from file
The main issue I had with regards to the slow execution turned out to be
related to how I was dealing with adding keyframes to the list. I was
ading four times as many as I thought I was which was slowing everything
down a lot. When I changed things I made it so the if statement was
never true (x - 1 % 9 == 0) for larger x which gave an impressively fast
runtime. Now I am adding a keyframe every 39 steps, in line with the
computation from hoomd and have fixed the issue with the modulo (order
of operations is important).