Getting positions working correctly
The positions are stored as a vector of the relative positions of the
particles, i.e. they vectors in the basis set of the unit cell
parameters. There are functions set_positions and get_positions which
will set the positions as well as get the positions in a format suitable
for using in hoomd.