Return to basics for repository
Removing most of the complicated scripts which are messy and hard to
understand so I can rebuild the functionality from the ground up with
some more sense. I am also using Jupyter notebooks for running the test
set of code since this is far easier to iterate with and to create
visualisations.