Move to pandas dataframes for data transport
Now using pandas data frames for moving the translational and rotational
data between functions since this is far simpler than the method I was
using before and far less prone to weird errors. There is also the
additional advantage that it is incredibly simple to then save the data
to an hdf5 file which is also beneficial.