Return dynamics quantities as a pandas DataFrame
The structure in which I was saving the data previously was horrible to
deal with and query. By using a DataFrame I have a rich array of tools
to both query the data and to simply visualise it. There is also the
advantage that I can easily save the DataFrame to an hdf5 file for later
querying.