Write hdf5 file with dynamics quantities
I am saving all the dynamics quantities to a hdf5 file which makes it
easy to come back to and perform the processing later. By having
everything in a pandas data frame it is simple to perform tasks like
taking the mean of the columns to then plot the data.