Optimise generate_steps function
Both optimising the generation of steps by using a simple python
generator. I have also debugged the generator so that it no longer has
off by one errors.

Along with the optimisations of the code I have also used the numba it
which shows small optimisations to the speed in tests generating a list
from the generator. I don't actually know if this will have real work
implications, but it is a simple change.