Use tempfile to save modifications
Using a NamedTemporaryFile for to save a gsd file with the changes made
in randomising the momenta. This is because the gsd_snapshot command
doesn't work for some reason, instead raising a segmentation fault.

The snapshot from gsd.hoomd doesn't play nice with the
hoomd.data.snapshot class, meaning I have to read in from a gsd file,
make the modifications, save to tempfile then read that tempfile in to
get the initial conditions.