Create functions for initialisation of configuration
I have created a series of functions that will initialise the simulation
from a snapshot, a file, or from nothing. The 'base' function is the
initialisation from a snapshot, which is used by all the other functions
in their initialisation.

I have also generated a series of tests for the initialise functions, as
well as for any bugs that I found along the way in any of the other
modules. The majority of the testing is really simple, but the idea is
that when I find the cases where things break I can develop a test for
them so it doesn't occur again next time I change something.