Include data analysis module
I have included a data analysis module in this repository since the
analysis of the data is linked to the generation of the simulation data.
Currently there is very little content in the data analysis module,
however I have made significant modifications to the TimeDepMany class
which allows for the offline analysis of the dynamics. This should also
be much more stable as I am outputting intermediate results to an hdf5
file.