Add more options to run_simulation
I have included a few more options that can be passed on the command
line, along with a simpler way of continuing to add options. I am using
the __dict___ property of the namespace to pass the values, so all the
options that have been set will be passed to the function to run.