Bugfix - create single additional generator/timestep
The GenerateStepSeries function was adding a new generator to the list
of generators every time the step was encountered, which exponentially
became a larger problem. This is now fixed with an additional set to
which steps are added when a generator is created at that step.

Along with this I have used the logging interface to create logs for
debugging the code which are currently turned off.