Datacolumns in hdf5 files
This modifies the code such that the haft files output are indexed by
the time and start_index columns, allowing for on disk queries of these
columns. This is incredibly useful for when the data doesn't entirely
fit in memory.

Additionally to remove any confusion I have renamed the column that was
index to start_index. This was an issues since the column by which the
file is indexed by is also referred to as index.