Creating a conda command line tool sdrun
Creating and installing a command line tool that allows me to call the
program from anywhere. This will also, after uploading to somewhere,
allow me to be even more effective at version control. Currently I am
using a local path rather than that of the git repository which is a
requirement of the current learning and hence instability of everything.
I want to be able to test and iterate as quickly as possible.