Create command line option for creating crystals
I have created a subcommand which allows for running different functions
based on the subcommand. This is great since it allows many different
functionalities from a single interface.

The one issue is that there is no really simple way to inherit options,
I have to specify them somewhere and then apply them to the function.
This is something that is going to require some understanding and will
therefore be rather flexible.