Subcommands recognised by sdrun
There was some issue where the subcommands of sdrun were not recognised
by the command group. It seems to work now, possibly a result of not
having parentheses when adding the @main.command() decorator.

Another change is that I am replacing main with sdrun for more
consistency in naming things.