Make arguments explicit
I had been passing kwargs to a number of functions, most of them
actually. The problem with this is that it hid the arguments that were
important to the function, the arguments it needed. It became difficult
to work out what something did.

So to improve the readability and reliability of my code I have switched
to passing all arguments explicitly and with that I am also typing all
the arguments as I go as well.