Support creation of liquid/crystal interface
I have developed a method to create a liquid crystal interface in the
trimer simulations. This involves heating up the entire simulation box
in an nvt environment so that I can scale the size increase of the
crystal with that of the liquid. After researching some more of the
capabilities of hoomd this is actually a silly approach, but it works as
a first approach.