Migrate to NPT simulation for all equilibrations
There I have changed the equilibration to now use npt simulations
instead of a series of NVT simulations. This also means that I am now
using the same code for both the equilibration of the crystal as the
melting of the interface, the only difference being the group which this
is applied to.