Output progress data for equilibration
I found out that I had no data output during the equilibration runs
which makes it very difficult to know what is going on. The output is
to the same directory as the output file, however this may have to
change somewhere down the line.

I am also migrating the set_dump and set_thermo commands to take a
filename rather than a pressure and temperature. This makes them a
little easier to use, and I can define the filename in a single
location. I say I am migrating as only the new calls in the
equilibration module have been changed at this point.