Remove equilibration of crystal in interface
Running the simulation to equilibrate the interface ended up melting the
crystal before the melting phase of the simulation. Since this is a
rather large problem I have completely removed the equilibration of the
crystal phase, assuming that the compression of the liquid phase will be
close enough to appropriate for the crystal phase.