Find memory leak issue and bolster test suite
The issue is really only manifesting as a memory leak, the real problem
is that the crystal lattice was trying to shear and find a lower energy
state. With an orthorhombic cell this wasn't allowed to occur properly
due to the boundary conditions. This weird behaviour at the boundary was
responsible for the gigantic memory useage, although I am not sure why
it was getting so large.

In fixing this issue I am using hypothesis far more for the test suite,
which should mean that the problem is more likely to be observed if it
pops up again. I am also increasing the number of steps for the
simulations in the tests again since I am running them more deliberately
now.