Manually run garbage collection after each simulation_test
This was an issue that arose when running the test suite using
hypothesis, starting many simulations in a short space of time. After
reading a thread on the hoomd mailing list https://groups.google.com/forum/#!topic/hoomd-users/wWtZm5k8C_s
it appears that the python garbage collector is not acting fast enough
in cleaning up links to open files. To prevent this occurring I have
added a manual gc.collect() at the end of each simulation test so this
problem doesn't show up again.