Compute order parameter and plot on figure
I have added a function to compute the orientational order parameter,
which allows for a rudimentary analysis of the regions that are
crystalline or not. I have started with implementing this for a figure
since this is something I really need to see to be able to set the
threshold for crystallinity.

This also includes a rather extensive redesign of the way I generate the
figures, instead of plotting every atom, I am now computing the
positions of the additional atoms in the molecule from the orientation.
This allows me to skip the slow cleanup step where I had to ensure all
the particles had the appropriate orientation for the molecule.