Create analysis module
I have created an analysis module for all the analysis code. This means
the analyses that I had used in the creation of the figures has moved to
the analysis module.

Currently this is very basic, however I plan to include all the
different analyses that I develop to be runnable from this single
location.