Change default dump output to rigid_center
By changing the output to only include the central particles I am
reducing the size of the output file by 1/3. It also means there is less
work to be done processing the file, since I am not computing any of the
values for the other parts of the rigid molecule.