Open single gsd file when outputting trajectory
There was an issue where I was opening far too many files to write to
when trying to output a hoomd frame. Rather than open a file and
continually append to it, I was opening a file every time I wanted to
write. This ran into an issue in that it reached the maximum number of
open files supported by the OS as the python garbage collector was not
fast enough closing all the files.

To solve this I am opening a single file which is written to using the
`write_restart()` function of the gsd file handle.