Add pg and p2gg crystal phases, update p2
This is adding new crystals for the trimer so that I am able to
calculate the melting point of these other crystal phases. This has been
tested to look like the images from our collaboration with Toby. However
they are not exact, which is totally fine since we are using a potential
rather than hard discs which does change the optimal positions somewhat.