Support disc class in simulations
I was assuming that I would always be using a molecule with a
rigid_center group defined, however this is not the case with the disc
or the sphere classes. I have changed instances where I am using a group
that is not `all` to check for the number of particles in the molecule.