Tests with updated p2 crystal pass
The updated p2 crystal I am using has a _magic number_ that is a
multiple of 6. Unit cell lengths that are a multiple of 6 work while
others result in errors. This is a result of the way I am generating an
orthorhombic unit cell which is a complete hack, but it does make the
resulting analysis simpler.