Generalised functions for computing dynamics
The functions I currently have are very dependent on the format of the
hoomd simulation. I am generating a series of more general functions
that only require the input of numpy arrays. This level of abstraction
makes the set of functions far more applicable to someone other than me,
and it is also easily transferable.

I have only implemented the simple functions at this stage, however
it will be relatively straightforward to implement further
functionality. The base functions `rotationalDisplacement` and
`squaredDisplacment` are functions that have been optimised and
benchmarked, with these versions being up to 10 times faster than
alternatives using `numba.jit`. There are further alternatives for
speeding up these function calls like cython that I can explore. The
idea is that it will be straightforward to implement these optimisations
in just the single functions that require them.