Add pressure as a command line argument
I want to be able to run a disc simulation which uses a much lower
pressure than what I am using for these molecules. Passing the pressure
on the command line is a simple change which gives me this flexibility.
Most of the code was already in place, I just had to pass it as an
argument.