Sensible number of steps between keyframes in run_npt
The number of steps between key frames was 1000, which is far too few to
be indicative of a new configuration. I have chosen 20000 as the number
of steps as this is long enough that the starting configurations are
likely distinct. I don't need completely new configurations, i.e. longer
than the relaxation time so this is a reasonable length of time.