Performance and optimisation of rotaional distance
The rotational distance formula was both fragile in the input it would
handle appropriately in addition to leaving a significant amount of
speed on the table. The original implementation took about 250 us for
an array of 1000 elements. Removing the error handling code reduced that
to ~80 us. The implementation I have now reduced that number even
further to ~60 us, for an overall reduction of about 4x in performance.