Make StepSize module more robust and succinct
The StepSize module was a somewhat hacked together codebase with
functionality essentially defined by the code rather than what I wanted
it to do. This update dramatically simplifies the code, especially of
the GenerateStepSeries class, and makes much better use of the available
generators.

This commit is focussed on the cleanup of the codebase. There is still a
fair amount of work to be done on the documentation and the testing of
the code.