Bugfix: No excess calls to the _enqueue function
The logic I had set out in the GenerateStepSeries class meant that I was
trying to add functions to the queue far too many times. This resulted
in slow runtimes for the function.

Now I have fixed this logic I am able to run the hypothesis tests over a
much wider range of values, so I am also opening up the range of values
for which I run test cases.