Change semiprivate class variables in GenerateStepSeries
I am exposing curr_step as a property of the class and making the queue
datastructure semiprivate. It makes far more sense this way since the
queue shouldn't be touched except through the methods of the class while
the current step is actually something that is rather useful to query.