Update plotting of figures and interactive_config
The interactive_config was using a different method of plotting the
molecules than the plot function I have in the configuration module.
I am now using the same functions to handle both these cases. In
addition I have made the functions more generalised so that the molecule
I am using to plot with defines the position of the additional
particles.

Using the same logic for both the bokeh server and the plot function
will make it much simpler to optimise the performance, and to add new
functionality in a single location.