Allow for equilibration of crystal phase
I have added additional functionality to the equilibration command which
will allow me to pass options that modify the output interval of the
dump files, in addition to running the equilibration of a crystal phase.
The equilibration of the crystal phase is important as it allows the
unit cell deform in a more flexible manner.