Buffered writing of analysis results
The results from the dynamics analysis were too large to fit in memory
in a number of cases. So I have included the option of passing an output
file to the read script which when present will write the results to the
file. The writing is done in a buffered fashion to optimise performance.

One decision which hurts performance was the choice to not hold the file
object for the duration of the analysis. This was chosen for both
simplicity of the code, but possibly more importantly being able to have
multiple threads writing to the same file as they release the file
handle.