Function to compute dynamics from trans and rot
The translation and the rotation of each molecule is the fundamental
building block of all the dynamics quantities I am calculating. This new
function computes a series of dynamics values for the array of input
translations and rotations. The idea being that these values are for a
single configuration, although multiple configurations would also work
just fine.