Remove Intermediate Translations and Rotation step
The processing step where I was converting a series of gad files to
translations and rotations was not working effectively. This
processing was essentially reformatting the gsd file into something
I that was even harder to deal with and take longer to process.

Instead I am going straight from the gsd files to dynamics quantities,
in a pandas DataFrame. This is great since it keeps the bulk of the
processing in a single step, the remaining steps should be fast and
straightforward.

Also included in this update is a series of dynamic quantities that are
being computed. Each is a function which allows for simple optimisation,
documentation and maintenance.