Bugfix - squaredDisplacement function periodic distance
The squaredDisplacement function was not computing the periodic
distances properly because I was trying to be too clever using the
square of the values rather than doing multiple comparisons.

I have now solved the issue, using code that is efficient. Noting that
the numpy implementation is the fastest I have generated. I have also
created a simple and slow reference implementation to use for testing,
which with hypothesis offers comparisons across a wide range of values.