Make displacement base quantity
Previously I was using the squared displacement as a base quantity for
the dynamics analysis, mostly with the idea that it was faster to
compute with the sqrt operation being slow. The `np.linalg.norm`
function is fast enough that it doesn't really make a difference, and it
is also much more readable having displacement rather than
squaredDisplacement. This also keeps it in line with the rotational
displacement.