Include moment_inertia_scale as parameter
One of the parameters that we want to be testing is the effect of the
moment of inertia on the dynamics. This adds a moment_inertia_scale
parameter to the trimer molecule which allows for dynamically modifying
the moment of inertia.

Along with the change to the molecule I have also allowed for a command
line parameter to modify the moment of inertia. This is passed to all
the functions which run the simulations.