Rename molecule.py to molecules.py
I have renamed the molecule module to enable more consistency with
naming things. I had some instances where I was using `mol` as the name
of the molecule instance while others were using `molecule`. So to
remove any issues of name collisions and to make everything simpler I
have changed the name of the molecule.py module to molecules.py.

This now means that I have
- molecule -> the instance of the molecule
- molecules -> the molecules module from which the different molecules
are imported
- Molecule -> the base molecule class

At the very least this enables consistency of usage throughout the
project.