Bugfix - Remove default step limit from analysis
I had issues with the analysis of the trajectories where the analysis
was only running for the first 1000 steps rather than all the steps.
This was a result of re-using the steps option from the running of the
simulation, which was set to a sensible default value of 1000. This is
definitely not sensible for the analysis so I have created a new option,
`step-limit` which has no limit by default.