Pass SimulationParams class holding arguments
There are a lot of arguments being passed around to define how to deal
with each simulation. Instead of passing around and dealing with the
arguments individually I have created a SimulationParams class which
will hold all the arguments and the logic of them in a single location.

This should also make it easier to understand the function calls and
remove issues of setting defaults in many different locations.