Use argparse for argument parsing to sim_params
Click wasn't able to do what I wanted with the command line arguments so
I have now migrated all the argument parsing to argparse. This is also
nice since it removes a dependency from the codebase.

This change is also beginning the migration of all functions to use the
SimulationParams to pass most of the arguments, so now all the command
line functions are using the sim_params.

These is still a lot to do to get this working and tested however because
I am now using argparse it should be a lot simpler to test that the
parsing of arguments occurs as I expect and that the appropriate vales
are changed in the sim_params class.