Setup params to properly handle attributes
The SimulationParams class now handles both the attributes it is set up
with, but also any attributes that are added by way of being in the
dictionary that holds all the values.

This is the most sensible method of holding the data since it is
flexible, it allows for the setting of default values and I can easily
customise the handling of both the setting and getting of specific
attributes.