Reimplement verbose to logging
In the migration from click to argparse the setting of the logging flags
was not implemented. This implements the ability to toggle the logging
level interactively on the command line.

All the code that was in options has been moved to the main.py class
which simplifies everything a little. Most of the code had been moved
previously, this moves the last of the code that was hanging around in
that module.