Bugfix - The neighbour list has no persistence
This was an interesting bug to fix which is mostly the result of some
weirdness due to the interfacing of C and python.

I was returning the result of nn.getNeighborList() which is a pointer to
an array. Once the NearestNeighbours class gets destroyed the memory is
free to be modified by other parts of the program. The bug I was running
into was a result of this, memory at the start of the array being
written to by some other part of the process causing unusual results.
I have solved this issue by copying the values of the array to a numpy
array which can then be returned as a python object with its own
lifetime.