Bugfix - Correct calculation of angle from quaternion
This fixes the calculation of the z angle from a quaternion to be
correct in all the quadrants, rather than just 2 of them. I have also
included tests to prevent regression of this in the form of specific
values.

This change also means that the quaternion2z will also now work for
quaternions that have rotations in other planes, giving the correct
angle of rotation in the z plane.