Bugfix - computation of dynamics
This fixes a couple of issues with regards to the computation of the
dynamics quantities. The first is that I was skipping the first frame of
the trajectory, incrementing the step counter before processing the
first frame.

I have also fixed an issue where the result arrays were being passed as
float64 values while the functions to compute them were expecting
float32 values. I have now created a single location for which the
values can be changed.