Script for generating pbs submission files
The main source of errors in my simulations is the generation of the pbs
scripts. By using python to generate them this eliminates many of my
errors. At this point there is just the generation for a single
simulation type, however this will eventually evolve into something a
whole lot more useful.