Use crystal in naming whenever passed
There are occasions like when running simulations of the interface
between the liquid and the crystal where although the crystal has
already been defined, for the purposes of identifying files it would be
useful to have as a component of the filename.

This adds the crystal structure as a component of the filename whenever
it is passed and allows for it to be passed in any of the simulations.