Add file to generate a series of interfaces
This new file interface.py generates a file for submission to a pbs
scheduler as an array job. This handles most of the work of generating
the script, there is just the variables to be created.

Currently this is a very rudimentary implementation of this idea. Once I
have a few of these scripts I plan to combine everything into a function
which is a single interface for everything.