Create interface with completely surrounded box.
I am currently interesting in the melting of any interface, not one in
particular. To make some of the issues I am having with the periodicity
of the crystal surface I am moving to a completely surrounded crystal,
which has the added benefit of being able to compare melting across
both faces at the same time.