Compute number of voronoi neighbours
Using distance to compute neighbours is a rather arbitrary process, so
for the calculation of order using the number of neighbours I have opted
to use the voronoi cell for the number of neighbours.

Currently this calculation is being done using the package tess, which
is a cython wrapper around voro++. This approach is too slow for what I
would like with my interactive visualisation, so I will likely look to
optimise it.