Functions to compute displacement return arrays
Rather than having the functions that compute the displacements take the
result array as an input value, they will now return the array. This is
a move to make these functions easier to use, and since I am using them
to interface with cython functions to actually do the computation it
makes sense for them to handle the creation of the arrays.