Update testing of quaternion maths
I am updating the testing to use the quaternion package for the
reference implementation of the calculation of rotations.  This ensures
that I am not making any mistakes in my assumptions about the
quaternions. The reason this package is not used throughout is that it
is slow for the types of computations I want to use it for.