Specify versions of dependencies
There is an issue on macOS with the latest version of the mkl/numpy
stack. I believe the issue has to do with hoomd, although I am currently
unsure, with only an abort raised. The latest versions work fine with
linux, so I believe it is just a matter of waiting for an update. In the
meantime I have specified the versions that will work on macOS.