Compute structural relaxation
I have added support for the computation of the structural relaxation
function. Since I am using the position and the orientation of the
central molecule, I have developed the functionality to compute the
motion of each of the individual particles from the position and
orientation of the central particles.

The computation of the structural relaxation function has been
implemented in both the macroscopic calculations, in addition to the
microscopic individual particles.