Bugfix - Problem with using memoryViews
The cython memory views have a bug for arrays which can't be changed. In
my situation I believe it manifests in the arrays which hold the
orientations.

Currently the solution to this issue is to use the numpy arrays instead
which has drawbacks like requiring the gil, however it works.