More stable processing of gsd files
This accounts for frames of a gsd file being corrupted, which is
something that is actually reasonably common in my workflow. This is a
result of copying the files while they are being written to, however
while this might be a bad idea it also means I am then free to process
the files.