Pass dictionaries for computed quantities
I was passing pandas DataFrames for the intermediate values, of the
dynamics computations. This was really silly since the creation of a
pandas DataFrame is really computationally expensive. Instead it is
about 400 times faster to just create a single DataFrame from a list of
dictionaries rather than many DataFrames which are concatenated
together.

In my simulation code it is about 5 times faster to compute the dynamics
quantities now that I don't have to create the DataFrame to return the
values.