Average the structural relaxation time over molecule
To be consistent with the rest of the quantities I am calculating, I am
now averaging the structural relaxation time over the entire molecule.
Since all the other quantities are per molecule there is no loss of
granularity in comparison of the mean. There could be some minor changes
in the distribution though.

This change means that all the values will have the same length, which
allows for much simpler analysis not having to deal with the unusual
case. It should also reduce the storage footprint as well, not
requiring all the NaNs for the missing values.