Add tests for machine learning models
The machine learning models were lacking tests of working. I have added
these with a small degree of checking for some correctness, in that all
the configurations should return as crystalline.

The additions of this testing also involves the update of the
environment and meta.yaml files to be consistent with the packages
required for the machine learning aspect.