Significantly increase coverage and number of tests
This is a drive to ensure that there is sufficient coverage of all the
functionality of this package so that when things change in the future
there is an awareness of what has broken. This should remove much of the
issues with having the exactly correct version of the software to get
the same results.