Split MD-Molecules-Hoomd repository
The MD-Molecules-Hoomd repository was getting too unwieldy, having code
from the running of simulations to the analysis. While this was great
for being able to change a bunch of things quickly, now it is currently
hurting me for requiring such a large web of dependencies. This also
makes it difficult to distribute the code, since that also requires a
huge number of downloads to go with it.

To break up the code in a sensible way I now have an sdanalysis side,
which holds all the scripts for the analysis, while there sdrun side
will handle the running of the simulations.