Replace freud with scipy for nearest neighbours
I believe the issue I have been having with the errors on the Mac has
been that the freud build won't link to the python 3.6.3 libraries for
some reason. So I have removed the dependency, replaced the nearest
neighbours calculations with a KDTree from the Scipy library and updated
all the code. With this update I have also removed most of the code in
the ordering from cython, changing it back to python since it was just
too unwieldy using cython for those small sections of code which were
mostly python code anyway.