Improve handling of subprocesses
The handling of the subprocesses spawning a bokeh server was done
poorly, this is a better implementation, leaving bokeh to handle the
SIGINT. I have also removed the test for running the figure as it left
an orphaned process I was unable to prevent occurring. This caused
unexpected behaviour on the second run of the tests.