Better specification of versions in meta.yaml
This uses the more specific versioning available, also pinning the
python and numpy versions to those of the build environment. This also
has less restrictive versions for all the run dependencies.

In specifying the versioning I have included a conda_build_config file
to potentially build for other versions of python/numpy although at this
stage it is not required.