Updates to travis build
This includes a series of semantic changes to make it simpler to
understand. All the setup of the build/install environment is performed
at the before_* stages.

This update also includes macOSX as run target once again. Now
understanding that python doesn't work properly I am using the generic
language which should work for both osx and linux.