Restructure tests using fixtures
The pytest fixture is a flexible method of defining parameters with a
given lifetime which is simple to modify. This is far more flexible than
just using global variables. I have changed a number of the tests to
using module level fixtures rather than module defined variables for the
flexibility it provides. There is also excellent support for using a
context manager allowing for file management in a single location.