Use pipenv for testing and deployment
This is a change due to the complexity of trying to get the conda
environments working with the makefile. This is a much simpler
alternative using pipenv to manage the virtual environment.

This does have the downside that it is not deployable using conda at
this stage, it is possible to do this although a little more
complicated.