Add custom versioning
The versioning scheme was picking up the extra files installed in the
directory when building on travis-ci. This meant that there was a local
(dirty) tag being appended to the version number. Since PyPI doesn't
accept this local part of the version number I have had to create a new
function to always ignore the local part of the version number.