Create manifest file with files to include
When generating an sdist there were files that were not included which
were required to build the package. The one that was primarily raising
an error was the voro++.hh file, however this was only one of the
missing files.

Since these were not python files, and not referenced in the setup.py
script, most of the voro directory was not included in the distribution.
By adding the files to the MANIFEST.in I am forcing their inclusion.