Break analysis of file into smaller chunks
Rather than just having a single function to read and analyse an input
file, I have broken down the analysis into a number of parts. The main
part of this decomposition is the movement of the iteration to a
generator which performs the iteration over the steps, and the frames of
the file. This makes the actual computation of the dynamics quantities
significantly clearer.

The additional change is the migration of saving the results to file is
now managed by a separate class. This writes the results to a file when
the list gets large enough.